fixed a few compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7156 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ features/include/pcl/features/impl/3dsc.hpp
@@ -195,7 +195,7 @@ pcl::ShapeContext3DEstimation<PointInT, PointNT, PointOutT>::computePoint (
 
     /// ----- Compute current neighbour polar coordinates -----
     /// Get distance between the neighbour and the origin
-    float r = sqrt (nn_dists[ne]);
+    float r = sqrtf (nn_dists[ne]);
 
     /// Project point into the tangent plane
     Eigen::Vector3f proj;
